+++ packages/lib/components/toast/types.ts
@@ -1,0 +1,9 @@
+export type ToastConfig = {
+  body: string;
+  duration?: number;
+  severity?: 'info' | 'success';
+};
+
+export type ToastWithId = Required<ToastConfig> & {
+  id: number;
+};